Add web/WASM platform support, fix lints, prep 2.0.0 release
- Add cross-platform support (web/WASM) via conditional exports for file
  state storage, HTTP adapters, and JavaScript runtime
- FileStateStorage now takes a String directoryPath instead of Directory
- Surface network errors via DioException instead of dart:io SocketException
- Add Platform Support section with web-safe API key snippet to both READMEs
- Fix README example references (remove kimi_vision, add Bedrock Claude)
- Style cleanup: lowerCamelCase identifiers and braced flow-control (pana 160/160)
- Bump version to 2.0.0, add issue_tracker, add AGENTS.md
