+++ pubspec.yaml
@@ -1,8 +1,9 @@
 name: dart_agent_core
 description: A mobile-first, local-first Dart library for building and evaluating stateful, tool-using AI agents with multi-provider LLM support.
-version: 1.0.14
+version: 2.0.0
 repository: https://github.com/memex-lab/dart_agent_core
 homepage: https://github.com/memex-lab/dart_agent_core
+issue_tracker: https://github.com/memex-lab/dart_agent_core/issues
 topics:
   - ai
   - agent
